rasp/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-bcrypt==1.0.1
flask-migrate==4.1.0
flask-cors==6.0.2
marshmallow==3.26.1
python-dotenv==1.0.1